requirements.txt
praw
asyncpraw
pymongo
pandas
matplotlib
python-dotenv
flask
flask-cors
vaderSentiment
textblob
newsapi-python
pytrends
yfinance
requests
schedule
nltk
lxml
html5lib
beautifulsoup4
